Initial upload of Imperial model
This model was provided by Swapnil Mishra (https://github.com/s-mishra/bench-marking-samplers-covid) to benchmark with an SMC sampler. As such, it is no attached to any particular paper authored by imperial. Uploading this as a base example of an imperial model as it contains fewer extra files than their main repository. The Stan model is largely the same as other imperial models but in this version, they attempted to remove several funnels. We may wish to switch to a modern version at a later date i.e. from https://github.com/ImperialCollegeLondon/covid19model